+++ lib/src/settings/asset-resolver.dart
@@ -31,32 +31,37 @@ class AssetResolver {
         assetUrl, _configuration.stillSizes, _qualitySettings.stillQuality);
   }
 
-  _getPathWithResolution(
-      String contentUrl, List<String> sizes, AssetQuality resolution) {
-    if (contentUrl == null || contentUrl.isEmpty) {
-      return null;
-    }
+  String _getPathWithResolution(
+    String contentUrl,
+    List<String> sizes,
+    AssetQuality resolution,
+  ) {
+    if (contentUrl == null || contentUrl.isEmpty) return null;
 
+    String sizeUrl;
     switch (resolution) {
-      case AssetQuality.Highest:
-        return _getPathWithSize(contentUrl, sizes, sizes.length - 1);
-      case AssetQuality.Lowest:
-        return _getPathWithSize(contentUrl, sizes, 0);
+      case AssetQuality.Original:
+        sizeUrl = sizes[sizes.length - 1];
+        break;
+      case AssetQuality.High:
+        sizeUrl = sizes[(sizes.length - 2).clamp(0, sizes.length)];
+        break;
+      case AssetQuality.Low:
+        sizeUrl = sizes[0];
+        break;
       case AssetQuality.Mid:
-        return _getPathWithSize(contentUrl, sizes, (sizes.length / 2).round());
+      default:
+        sizeUrl =
+            sizes[((sizes.length - 2) / 2).round().clamp(0, sizes.length)];
     }
-  }
 
-  _getPathWithSize(String contentUrl, List<String> sizes, int size) {
-    final clampedSize = size.clamp(0, sizes.length - 1);
-
-    final sizeUrl = sizes[clampedSize];
     return "${_configuration.secureBaseUrl}$sizeUrl$contentUrl";
   }
 }
 
 enum AssetQuality {
-  Lowest,
+  Low,
   Mid,
-  Highest,
+  High,
+  Original,
 }