+++ lib/src/settings/movie-search-settings.dart
@@ -1,26 +1,30 @@
+import '../models/date.dart';
+import '../models/sort-by.dart';
+
 import 'quality-settings.dart';
 
-class MovieSearchSettings {
-  String language;
-  String query;
-  int page;
-  bool includeAdult;
-  String region;
-  int year;
-  int primaryReleaseYear;
-  QualitySettings quality;
+part 'movie-discover-settings.dart';
 
-  MovieSearchSettings(
-      {this.query,
-      this.language = "en-US",
-      this.page = 1,
-      this.includeAdult = false,
-      this.region,
-      this.year,
-      this.primaryReleaseYear,
-      QualitySettings quality}) {
-    this.quality = quality ?? QualitySettings();
-  }
+class MovieSearchSettings {
+  final String language;
+  final String query;
+  final int page;
+  final bool includeAdult;
+  final String region;
+  final int year;
+  final int primaryReleaseYear;
+  final QualitySettings quality;
+
+  const MovieSearchSettings({
+    this.query,
+    this.language = "en-US",
+    this.page = 1,
+    this.includeAdult = false,
+    this.region,
+    this.year,
+    this.primaryReleaseYear,
+    this.quality = const QualitySettings(),
+  }) : assert(quality != null);
 
   Map<String, dynamic> toJson() {
     var map = Map<String, dynamic>();