+++ lib/src/settings/quality-settings.dart
@@ -1,16 +1,17 @@
 import 'asset-resolver.dart';
 
 class QualitySettings {
-  AssetQuality posterQuality;
-  AssetQuality backdropQuality;
-  AssetQuality logoQuality;
-  AssetQuality profileQuality;
-  AssetQuality stillQuality;
+  final AssetQuality posterQuality;
+  final AssetQuality backdropQuality;
+  final AssetQuality logoQuality;
+  final AssetQuality profileQuality;
+  final AssetQuality stillQuality;
 
-  QualitySettings(
-      {this.posterQuality = AssetQuality.Mid,
-      this.backdropQuality = AssetQuality.Mid,
-      this.logoQuality = AssetQuality.Mid,
-      this.profileQuality = AssetQuality.Mid,
-      this.stillQuality = AssetQuality.Mid});
+  const QualitySettings({
+    this.posterQuality = AssetQuality.Mid,
+    this.backdropQuality = AssetQuality.Mid,
+    this.logoQuality = AssetQuality.Mid,
+    this.profileQuality = AssetQuality.Mid,
+    this.stillQuality = AssetQuality.Mid,
+  });
 }